# Improve the performance of indirect jump for T1C